Add an option to toggle eliding constant nodes (#2521)
Editing this node in a real time MaterialX editing software leads to constant shader recompilation because the value gets burned inside the shader text. On complex shading models, this has a noticeable performance impact.

Add an option to keep these as regular nodes that do not require recompiling the shader.