Fix crash on selecting functional graph output in Graph Editor (#2597)
Closes: https://github.com/AcademySoftwareFoundation/MaterialX/issues/2428

### Issue
There was an issue where when you click on an output node in Functional Graphs in the MaterialX graph editor, it would crash.

### Fix
The reason for this was that output nodes return a nullptr on getNode() calls, and so were being missed by the check for that determines if the node was inside a functional graph. I've added this check for output nodes, and now this crash doesn't occur.