Create MaterialXGenHw module (#2565)
Following on from the great work in #2555, this PR consolidates the shader generation code for HW shading languages in to a new MaterialX module - MaterialXGenHw. 

This PR does not change any lines of actual code, only moves them around to different files, and updates include statements to reflect the new headers.

`HwShaderGenerator.h` has been split in to separate files, which reduces the coupling to some downstream modules to just the constants in some places.

The MaterialXRender directly links to this new module, and that may not be optimal. We may want to refactor things such that only MaterialXRenderHw links to this new module, leaving the MaterialXRender module agnostic of shading languages, but this is outside the scope of this PR. I welcome discussion/PRs for this work if people think it is important.

I did not attempt to refactor the python or javascript bindings to reflect this new module because I'm not super familiar with that code, and am not a regular user of the bindings. If it is important to have the bindings reflect the C++ modules then I would welcome a PR to reflect this change in the bindings. 

It should be noted that this does break the API. As noted by @niklasharrysson [here](https://github.com/AcademySoftwareFoundation/MaterialX/pull/2555#issuecomment-3297514089) we've already broken API compatibility.  I think it might be possible to provide some shim code to provide compatibility, if that is important for downstream consumers, but I propose we "wait and see" if that is required.

On a similar note, a number of unnecessary header includes were removed, they appeared to be copy/paste inclusions, so it should be harmless, but technically it could impact downstream consumers if they were dependent on them.