Initial nightly build (#2621)
This changelist adds an initial nightly build to the GitHub CI for MaterialX, including a single OpenImageIO build in the job matrix.  The extended functionality of the nightly build can also be triggered manually via dispatch, allowing developers to validate their work on demand in their fork of MaterialX.

This is intended as a first step towards the integration of more complex and time-intensive steps in our GitHub CI, including options such as MDL shader validation, OSL shader validation, and more extensive render validation.