Support Specular and Specular Color in Standard Surface to glTF PBR  (#2623)
closes #2581 

## Summary
This PR passthrough the specular weight and the specular color from standard surface to gltf_pbr. 

This translation will not produce completely matching result, even if specular and specular_color control similar aspects in both the shading models, there are some fundamental differences. It will successfully translate the specular weight, so if you have a material with roughness 0, and specular 0, it will be a diffuse material (similar to roughness 1). But the difference will be in the specular_color, both for dielectric materials (non-metals) and conductors (metals).