Fix `thin_walled` and `transmission_depth` in Standard Surface to glTF PBR (#2644)
Thickness in glTF PBR is used to describe how thick a mesh is. If it's thin walled, thickness is 0, otherwise it should be a value corresponding to a mesh, in a ray tracer, you will have that info when rendering, for a rasterizer you either bake it or use an arbitrary value (such as longest side of the bounding box).

Since Thickness is a mesh related value, we can't translate to the expected value, unless the material is considered `thin_walled`. Then it should be 0. I have updated the translation graph from Standard surface to set thickness to 0 if thin walled, otherwise 1 (which is better then nothing).

Previously `transmission_depth` was used for thickness, which do not correspond to glTF PBRs thickness. I have changed it to be used for `attenuation_distance` which correspond to `transmission_depth`, (with the caveat that in glTF PBR it should >0, as `base_color` is used to color the surface even for `thin_walled` transmissive materials).