Fix mx_shadow_platform.metal filename (#2651)
I made a mistake when addressing notes in #2607.  This file was missed when renaming the source files.

This file is being included [here](https://github.com/AcademySoftwareFoundation/MaterialX/blob/main/source/MaterialXGenMsl/MslShaderGenerator.cpp#L934).

I'm really surprised, and a little concerned that the unit tests didn't catch this.  We should investigate what is necessary to improve the coverage here.