Improve ShaderGenerator API (#2652)
In #2604 two new API calls for the `ShaderGenerator` class were added to allow customization of the creation of `ShaderNodeImplementation` objects for `NodeGraph` and `Implementation` elements respectively.

I think the API is cleaner, and more robust if we pass the actual implementation objects, and not the associated `NodeDef`.  In some cases these function have to re-acquire the implementation object from the `NodeDef`, but at the call sight we already have it.

#2604 has not been included in a MaterialX release yet - so I don't believe we have any problems changing the API.