Implement OSL command string code generator (#2603)
This PR implements a new OSL code generator that emits shader group command strings as described in #2575. The work is from a Dev Days collaboration at Imageworks between myself, @JoshuaSenouf, and @MrigankaArya.

The three main parts are a tool to generate .oso's from MaterialX nodedefs, a ShaderGenerator class that generates the command string, and tests that run the existing OSL testrender cases with the new codegen.

At Imageworks we would like to use this to investigate performance differences between the two code generation styles, and to explore mixing and matching our internal shaders with MaterialX's.