Add MDL Compiler dependency for genmdl tests (#2254)
Extend test for MaterialXGenMdl.

This commit extends source/MaterialXTest/MaterialXGenMdl as a test for
source/MaterialXGenMdl, plus the necessary work on the cmake scripts and github
actions to add the MDL SDK as a dependency.

New cmake variables:
- MATERIALX_MDL_SDK_DIR: if set and mdlc is found the test in
source/MaterialXTest/MaterialXGenMdl is extended to run mdlc on the generated
MDL files.
- MATERIALX_MDL_BINARY_MDLC: name of the mdlc binary, derived from MATERIALX_MDL_SDK_DIR

The CI tests use vcpkg to install the MDL SDK, but a manually built one also
works (alternative for developers).

Right now, the MDL SDK dependency is enabled only for Windows_VS2022_x64_Python313 in the extended build.