Update `FilePath::getFilesInDirectory` to allow getting all files (#2633)
While working on some other internal projects I found myself using the `FilePath` object, and bumped up against the fact that `getFilesInDirectory` is unable to query all the files in the provided directory.

This PR provides a simple extension to the existing code to allow accessing all the files in a directory if no extension is provided to match against.