Improve data library tests for OSL (#2656)
OSL Tests fail when nodes with namespaces are used. The root cause is the test tries to determine the nodename from the nodedef by stripping out "ND_" prefix. This change reworks the implementation to query the nodestring from the nodedef and ensure shadergen works for various input/output variations of the node.