Fix some obscure swizzle upgrade bugs (#2637)
Was working on upgrading a MaterialX material when it is written using UsdShade data model, and that required reviewing the current upgrade code to convert it to Python/USD. Also was working on a unit test, which forced me to put on my QA hat and start devising devious ways to fail the code.

This resulted in the following two fixes for extremely rare corner cases.