Add support for df_vulkan as MDL test renderer.
Wrapper code based on that for the OSL test renderer and intermediate code
from #2254.

The new CMake variable MATERIALX_MDL_BINARY_TESTRENDER contains the path of the
df_vulkan binary, and is derived from MATERIALX_MDL_SDK_DIR.

Use x64-windows-release triplet for the MDL SDK to save disk space on the
runners (and build time).