Improve OSL render testing (#2661)
Currently the render unit tests are constrained to a small set of the mtlx files in resources.  I think we should run as many as we can - but thats a discussion for another day.... 

Locally if I enable all of the `Materials/Examples`, and all of the `Materials/TestSuite` folder there are a number of failures for OSL - this is because the shader it outputting an incompatible type with what `testrender` is expecting.

While working on the OSL network shader generator, @chellmuth added a new osl node `setCi` that can accept a number of different inputs and output a `closure color`. 

This PR generalizes, extends and replaces that idea. 

I'm adding a new MaterialX node definition specific for OSL use - which is why I've placed it in the implementation specific file - but I'm open to other ideas there.  This new node is largely the same as the node Chris wrote, but instead of leveraging the `isconnected()` mechanism, it uses an input integer to select which input to export from the node.  This is because in the OSL shader generator we don't have OSL connections. 

Both OSL shader generators then use a generator option to control if this node should be added inline with the outputs.  This ensures that all outputs are converted to `closure color` which means they should all be renderable with `testrender`. 

The GLSL shader generate was updated to correctly emit a boolean value to the image, previously we would always emit black. 

As previously requested - clang-format has been run on all the files and unit tests have been updated to skip testing this node, as it is not a "real" MaterialX node.