Fix CMake build rule for resource installation (#2687)
## Issue

The post build step for MaterialXTest does no trigger so resources are not copied to the
build area on a build. 

This means developers must manually copy this over when they are changing any files affecting the resources folder. If the lack of copy is not noticed / done then it can get confusing as to why source changes are not reflected in the build tests. 

## Change

Add an explicit target for resource copy and make MaterialXTest dependeng on it.