Consolidate `toVec4` method in HW generators (#2702)
While looking at some of the GLSL vs MSL differences @bernardkwok raised in #2697. I noticed that the `GLSLShaderGenerator::toVec4()` and the `MSLShaderGenerator::toVec4()` have gotten out of sync. I think it makes sense to continue the consolidation work and move this to `HWShaderGenerator` to be more easily maintained.

I think there could be future work here to follow the pattern introduced in #2661, and instead of having the conversion shader code hard-coded in the shader generator - we could add a new node to the data library that performs the conversion and just update the shader generator to insert it when necessary.