Fix Incorrect implementation of SplitTB, RampTB and Ramp4 node types (#1758)
When the input value ('V' in MaterialX spec, texcoord.y in the GLSL and OSL code) is below the 'center' input value, then the output should be the bottom value ('valueb'), otherwise if it is above then the output should be the top value ('valuet').
Given a default 'center' value of 0.5, it means that a texcoord.y value of 0.0 should yield 'valueb' as the output, and a texcoord.y value of 1.0 should yield 'valuet' as the output.

The glsl and OSL code seem to do the opposite: mix in GLSL (and similarly in OSL) returns the first argument value when the third argument is 0.0.
The third argument here is a call to mx_aastep(center, texcoord.y) which internlly uses smoothstep.
For a default 'center' value of 0.5, a texcoord.y value of 0.0 yields 0.0 so the third argument to mix() is 0.0, meaning the first argument to mix() is returned. texcoord.y being 0.0 is below 'center' of 0.5 so the MaterialX spec states that 'valueb' should be returned, and yet the returned first argument to mix() is 'valuet'.
Similarly, when texcoord.y is 1.0 being above a 'center' value of 0.5 the MaterilX spec states that 'valuet' should be returned, and yet this glsl implementtion returns 'valueb' in this case.
The same applies to `ramptb` and `ramp4` nodes, and also in the MDL backend.
This commit fixes this issue at the source OSL, GLSL and MDL files for `splittb` and `ramptb` nodes, and the node-graph implementation for `ramp4` node.