Improvements to StandardNodes tables
Building upon the work in the new `comparenodes.py` script, this changelist makes several improvements to the tables in the StandardNodes specification, bringing the specification into closer alignment with the data libraries.

Using the output of `comparenodes.py` as a measure, this changelist reduces the number of differences between the StandardNodes specification and data libraries from 122 to 46.

Additionally, this changelist enables GitHub CI for changelists containing only edits to Markdown files, so that `comparenodes.py` is run on changes that contain only specification updates.