AOUSD formatting proposals for PBR Nodes (#2732)
This changelist contains an initial set of formatting proposals for the PBR Nodes specification, organized by the AOUSD Materials Working Group.

In addition to the formatting updates, details on the expectation of energy preservation for the specular BSDFs in MaterialX have been added, including a link to Emmanuel Turquin's 2019 paper.